+++ core/common/src/main/kotlin/com/alreadyoccupiedseat/core/extension/ColorExtension.kt
@@ -1,4 +1,4 @@
-package com.alreadyoccupiedseat.core.extention
+package com.alreadyoccupiedseat.core.extension
 
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.toArgb
