+++ core/common/src/main/kotlin/com/alreadyoccupiedseat/core/extension/KotlinExtension.kt
@@ -1,0 +1,4 @@
+package com.alreadyoccupiedseat.core.extension
+
+val String.Companion.EMPTY
+    get() = ""
