--- core/common/src/main/kotlin/com/alreadyoccupiedseat/core/extention/KotlinExtension.kt
@@ -1,4 +1,0 @@
-package com.alreadyoccupiedseat.core.extention
-
-val String.Companion.EMPTY
-    get() = ""
