+++ core/designsystem/src/main/java/com/alreadyoccupiedseat/designsystem/component/BottomNavItem.kt
@@ -19,7 +19,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 import com.alreadyoccupiedseat.Screens
 import com.alreadyoccupiedseat.designsystem.ShowpotColor
-import com.alreadyoccupiedseat.core.extention.conditional
+import com.alreadyoccupiedseat.core.extension.conditional
 import com.alreadyoccupiedseat.designsystem.typo.korean.ShowPotKoreanText_B1_SemiBold
 
 @Composable
