+++ core/designsystem/src/main/java/com/alreadyoccupiedseat/designsystem/component/ShowPotSearchBar.kt
@@ -17,8 +17,8 @@ import androidx.compose.ui.graphics.SolidColor
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import com.alreadyoccupiedseat.core.extention.EMPTY
-import com.alreadyoccupiedseat.core.extention.conditional
+import com.alreadyoccupiedseat.core.extension.EMPTY
+import com.alreadyoccupiedseat.core.extension.conditional
 import com.alreadyoccupiedseat.designsystem.R
 import com.alreadyoccupiedseat.designsystem.ShowPotTypography
 import com.alreadyoccupiedseat.designsystem.ShowpotColor
